+++ src/test/java/io/github/astrapi69/file/FileTestCase.java
@@ -28,11 +28,10 @@ import static org.testng.AssertJUnit.assertTrue;
 
 import java.io.File;
 
-import io.github.astrapi69.BaseTestCase;
-import io.github.astrapi69.file.FileExtensions;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 
+import io.github.astrapi69.BaseTestCase;
 import io.github.astrapi69.file.create.FileCreationState;
 import io.github.astrapi69.file.create.FileFactory;
 import io.github.astrapi69.file.delete.DeleteFileExtensions;